feat: 라우터 추가 및 컴포넌트 분리
