+++ src/components/index.js
@@ -1,0 +1,2 @@
+export { default as Header } from './Header';
+export { default as List } from './List';
