+++ src/constants/urls.js
@@ -1,0 +1,10 @@
+export const API_URL = {
+  github: 'https://api.github.com',
+  organization: 'facebook',
+  repositoty: 'react',
+};
+
+export const AD_URL = {
+  to: 'https://www.wanted.co.kr/',
+  img: 'https://image.wanted.co.kr/optimize?src=https%3A%2F%2Fstatic.wanted.co.kr%2Fimages%2Fuserweb%2Flogo_wanted_black.png&w=110&q=100',
+};
