+++ src/main.jsx
@@ -1,12 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import GlobalStyle from './styles/GlobalStyle';
 import App from './App';
-
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
-    <GlobalStyle />
     <App />
   </React.StrictMode>
 );
