+++ src/pages/index.js
@@ -1,0 +1,2 @@
+export { default as Root } from './Root';
+export { default as Main } from './Main';
